+++ pmd-core/src/main/java/net/sourceforge/pmd/util/log/internal/MessageReporterScope.java
@@ -2,9 +2,10 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.util.log;
+package net.sourceforge.pmd.util.log.internal;
 
 import net.sourceforge.pmd.annotation.InternalApi;
+import net.sourceforge.pmd.util.log.MessageReporter;
 
 /**
  * A logger that prefixes a scope name to log messages. Also keeps a
