+++ pmd-core/src/main/java/net/sourceforge/pmd/util/log/internal/NoopReporter.java
@@ -2,9 +2,10 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.util.log;
+package net.sourceforge.pmd.util.log.internal;
 
 import net.sourceforge.pmd.annotation.InternalApi;
+import net.sourceforge.pmd.util.log.MessageReporter;
 
 /**
  * A logger that ignores all messages.
