+++ pmd-core/src/test/java/net/sourceforge/pmd/lang/document/TestMessageReporter.java
@@ -6,7 +6,7 @@ package net.sourceforge.pmd.lang.document;
 
 import java.util.logging.Logger;
 
-import net.sourceforge.pmd.util.log.SimpleMessageReporter;
+import net.sourceforge.pmd.util.log.internal.SimpleMessageReporter;
 
 /**
  * @author Clément Fournier
